another day at 42

+++ ft_memcmp.c
@@ -6,12 +6,12 @@
 /*   By: hhakim <hhakim@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/01 15:37:21 by hhakim            #+#    #+#             */
-/*   Updated: 2022/11/09 16:22:40 by hhakim           ###   ########.fr       */
+/*   Updated: 2022/11/15 19:20:49 by hhakim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
-//compares first n chars for two strings(*s, *t, n) 
+//compares first n chars for two strings
 // return the differnce in first unequal char as int
 int	ft_memcmp(void *s, void *t, size_t n)
 {
@@ -23,7 +23,7 @@ int	ft_memcmp(void *s, void *t, size_t n)
 	while (n)
 	{
 		if (*s1 != *s2)
-			return (*s1 - *s2);
+			return ((int)(*s1 - *s2));
 		s1 ++;
 		s2 ++;
 		n --;
